requirements.txt
slack_bolt~=1.18.1
python-dotenv~=1.0.1
requests~=2.31.0
slack_sdk~=3.27.2
wordcloud~=1.9.3
matplotlib~=3.9.0